+++ libs/cosmos-sdk/x/auth/ibc-tx/alias.go
@@ -2,7 +2,7 @@ package ibc_tx
 
 import (
 	"github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/adapter"
-	ibccodec "github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/ibc-codec"
+	ibccodec "github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/pb-codec"
 )
 
 var (
