+++ libs/cosmos-sdk/x/auth/ibc-tx/internal/pb-codec/proto.go
@@ -3,7 +3,7 @@ package codec
 import (
 	codectypes "github.com/okex/exchain/libs/cosmos-sdk/codec/types"
 	"github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/ethsecp256k1"
-	ibckey "github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/ibc-key"
+	ibckey "github.com/okex/exchain/libs/cosmos-sdk/x/auth/ibc-tx/internal/secp256k1"
 
 	//"github.com/okex/exchain/libs/cosmos-sdk/crypto/keys/ed25519"
 	//"github.com/okex/exchain/libs/cosmos-sdk/crypto/keys/multisig"
